Export star view
